# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaperTrail API",
    "description": "Academic paper discovery and research assistant API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agent/ask": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Ask Agent",
        "operationId": "ask_agent_api_v1_agent_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/papers/search": {
      "get": {
        "tags": [
          "papers"
        ],
        "summary": "Search Papers",
        "operationId": "search_papers_api_v1_papers_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Papers Api V1 Papers Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/openalex/works/{work_id}": {
      "get": {
        "tags": [
          "openalex"
        ],
        "summary": "Get Openalex Work",
        "operationId": "get_openalex_work_api_v1_openalex_works__work_id__get",
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^W\\d+$",
              "title": "Work Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Openalex Work Api V1 Openalex Works  Work Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/openalex/authors/search": {
      "get": {
        "tags": [
          "openalex"
        ],
        "summary": "Search Openalex Authors",
        "operationId": "search_openalex_authors_api_v1_openalex_authors_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "minimum": 1,
              "default": 10,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Openalex Authors Api V1 Openalex Authors Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/openalex/authors/{author_id}": {
      "get": {
        "tags": [
          "openalex"
        ],
        "summary": "Get Openalex Author",
        "operationId": "get_openalex_author_api_v1_openalex_authors__author_id__get",
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^A\\d+$",
              "title": "Author Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Openalex Author Api V1 Openalex Authors  Author Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/openalex/authors/{author_id}/works": {
      "get": {
        "tags": [
          "openalex"
        ],
        "summary": "List Openalex Author Works",
        "operationId": "list_openalex_author_works_api_v1_openalex_authors__author_id__works_get",
        "parameters": [
          {
            "name": "author_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^A\\d+$",
              "title": "Author Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "minimum": 1,
              "default": 10,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Openalex Author Works Api V1 Openalex Authors  Author Id  Works Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/arxiv/search": {
      "get": {
        "tags": [
          "arxiv"
        ],
        "summary": "Search Arxiv",
        "operationId": "search_arxiv_api_v1_arxiv_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "title": "Q"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9975,
              "minimum": 0,
              "default": 0,
              "title": "Start"
            }
          },
          {
            "name": "max_results",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "minimum": 1,
              "default": 10,
              "title": "Max Results"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Arxiv Api V1 Arxiv Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/arxiv/works/{arxiv_id}": {
      "get": {
        "tags": [
          "arxiv"
        ],
        "summary": "Get Arxiv Work",
        "operationId": "get_arxiv_work_api_v1_arxiv_works__arxiv_id__get",
        "parameters": [
          {
            "name": "arxiv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 4,
              "maxLength": 64,
              "title": "Arxiv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Arxiv Work Api V1 Arxiv Works  Arxiv Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/arxiv": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Catalog Arxiv",
        "operationId": "list_catalog_arxiv_api_v1_catalog_arxiv_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Catalog Arxiv Api V1 Catalog Arxiv Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/arxiv/{arxiv_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Catalog Arxiv",
        "operationId": "get_catalog_arxiv_api_v1_catalog_arxiv__arxiv_id__get",
        "parameters": [
          {
            "name": "arxiv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 4,
              "maxLength": 64,
              "title": "Arxiv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Catalog Arxiv Api V1 Catalog Arxiv  Arxiv Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/catalog/papers": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Catalog Papers",
        "operationId": "list_catalog_papers_api_v1_catalog_papers_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Catalog Papers Api V1 Catalog Papers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentQuestion": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AgentQuestion"
      },
      "AgentResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "insufficient_evidence",
              "unverified_citations",
              "max_steps_exceeded",
              "tool_budget_exceeded",
              "deadline_exceeded",
              "model_disabled",
              "model_unavailable",
              "model_refusal",
              "invalid_model_output"
            ],
            "title": "Status"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "tool_calls": {
            "type": "integer",
            "title": "Tool Calls",
            "default": 0
          },
          "model_steps": {
            "type": "integer",
            "title": "Model Steps",
            "default": 0
          },
          "model_input_tokens": {
            "type": "integer",
            "title": "Model Input Tokens",
            "default": 0
          },
          "model_output_tokens": {
            "type": "integer",
            "title": "Model Output Tokens",
            "default": 0
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "duration_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duration Ms",
            "default": 0
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/AgentTraceEvent"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Trace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "answer"
        ],
        "title": "AgentResponse"
      },
...
    }
  }
}